Changing the base shape of images passed from (feature, batch_size) to (batch_size, feature).
It was really difficult for me to make my backpass to work. After some hours of research on it I realised that I was unnecessarily using transpose on A changing shape of vector X. So, removing A = X.T it was a lot of refactoring until it finally worked. Still weights are not correctly being updated but at least no errors in backpass.